Create a base class for the Icons
